requirements.txt
# Discord Bot Framework
discord.py>=2.3.2

# Database
peewee>=3.17.0

# Logging
rich>=13.7.0

# HTTP Requests (for currency conversion, etc.)
aiohttp>=3.9.1

# Environment Variables
python-dotenv>=1.0.0

# Date/Time utilities
python-dateutil>=2.8.2
